Rename stack to form stack
